:truck: extract episode tile layout
