+++ front/src/components/episode/tile/Info.jsx
